requirements.txt
groq
streamlit
python-dotenv
requests
plotly
